+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 - PR #1292 Implemented Bitwise binary ops AND, OR, XOR (&, |, ^)
 - PR #1235 Add GPU-accelerated Parquet Reader
 - PR #1335 Added local_dict arg in `DataFrame.query()`.
+- PR #1282 Add Series and DataFrame.describe()
 - PR #1381 Add DataFrame._get_numeric_data
 - PR #1396 Add DataFrame.drop method
 - PR #1412 Add DataFrame.pop()